Remove PHP 8.1 deprecations
The missing return types of the `ShortidType::convertToDatabaseValue()` and the `ShortidType::convertToPHPValue()` methods raised deprecation warnings with PHP 8.1, so I introduced them with a strict scope. Returning nothing from these methods does not seem like a good practice, so I returned null, when $value is empty instead. And `ShortidType::convertToDatabaseValue()` should IMHO do an explicit string cast instead of relying on an automatic type conversion later on. All that changes also allowed enabling the strict type for that file.

I also added inheritdoc annotations to mark the methods as overriding.